Change ClosingDocument.bill_sum from int to float

+++ federal_empl_program/models.py
@@ -508,7 +508,7 @@ class ClosingDocument(models.Model):
         "Тип документа", max_length=4, 
         choices=DOC_TYPES, blank=False, null=False
     )
-    bill_sum = models.IntegerField("Сумма оплаты", default=0)
+    bill_sum = models.FloatField("Сумма оплаты", default=0)
     is_paid = models.BooleanField("Оплаченно?", default=False)
     
     def doc_path(instance, filename):
